Use properties for Event.ok and Event.defused
New `Event.ok` and `Event.defused` properties wrap `Event._ok` and
`Event._defused` attributes. The key benefit is that doc strings now
can exist for `Event.ok` and `Event.defused`.

This is a performance-neutral change since all of the fast-paths still
use direct access to the `_ok` and `_defused` attributes.